adds (3:)  Replace MealRepository и InMemoryMealRepository

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -3,11 +3,12 @@ package ru.javawebinar.topjava.repository.inmemory;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepository implements MealRepository {
     private final Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -19,7 +20,9 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public Meal save(Meal meal) {
-        if (meal.isNew()) {
+        if(meal.getUserId() != SecurityUtil.authUserId())
+            return null;
+        if (meal.isNew() ) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
             return meal;
@@ -30,17 +33,23 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id) {
+        if(repository.get(id).getUserId() != SecurityUtil.authUserId())
+            return false;
         return repository.remove(id) != null;
     }
 
     @Override
     public Meal get(int id) {
+        if(repository.get(id).getUserId() != SecurityUtil.authUserId())
+            return null;
         return repository.get(id);
     }
 
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+        Collection<Meal> sortedList = (Collection<Meal>)repository.values().stream().sorted(Comparator.comparing(o -> o.getDateTime()));
+        Collections.reverse(Arrays.asList(sortedList.toArray()));
+        return sortedList;
     }
 }
